reorganized collect and correct
